+++ pkg/provider/models.go
@@ -1011,11 +1011,11 @@ type ArtifactContainerModel struct {
 }
 
 type ArtifactEnvironmentVariableModel struct {
-	Source       types.String `tfsdk:"source"`
-	Name         types.String `tfsdk:"name"`
-	Value        types.String `tfsdk:"value"`
-	CredentialID types.String `tfsdk:"credential_id"`
-	Key          types.String `tfsdk:"key"`
+	Source         types.String `tfsdk:"source"`
+	Name           types.String `tfsdk:"name"`
+	Value          types.String `tfsdk:"value"`
+	DrCredentialID types.String `tfsdk:"dr_credential_id"`
+	Key            types.String `tfsdk:"key"`
 }
 
 type ArtifactProbeConfigModel struct {
@@ -1062,8 +1062,8 @@ type WorkloadContainerOverrideModel struct {
 type WorkloadResourceAllocationModel struct {
 	CPU       types.Float64 `tfsdk:"cpu"`
 	GPU       types.Float64 `tfsdk:"gpu"`
-	GPUMemory types.Int64   `tfsdk:"gpu_memory"`
-	Memory    types.Int64   `tfsdk:"memory"`
+	GPUMemory types.String  `tfsdk:"gpu_memory"`
+	Memory    types.String  `tfsdk:"memory"`
 }
 
 type WorkloadAutoscalingModel struct {